Changed spacing parameter type from int32 to float.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@27319 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/kits/interface/layouter/ComplexLayouter.cpp
@@ -157,9 +157,9 @@ struct ComplexLayouter::SumItemBackup {
 
 
 // constructor
-ComplexLayouter::ComplexLayouter(int32 elementCount, int32 spacing)
+ComplexLayouter::ComplexLayouter(int32 elementCount, float spacing)
 	: fElementCount(elementCount),
-	  fSpacing(spacing),
+	  fSpacing((int32)spacing),
 	  fConstraints(new(nothrow) Constraint*[elementCount]),
 	  fWeights(new(nothrow) float[elementCount]),
 	  fSums(new(nothrow) SumItem[elementCount + 1]),
